Stop using exiftool as it's very slow compared to using native node.js methods
Also implemented a helper method for loading files which is async
instead of sync. This will also be handy for implementing "virtual
files".

Will bring it back as a separate plugin that people can opt into and
where it'll also read metadata only on demand so only encur cost when
people really need the extended file metadata.