Move old default starter to 'kitchen-sink' and replace with minimal starter (#806)
I've been hearing pretty consistent feedback from people trying Gatsby
that the old default starter is confusing as it was built to show off
many of the capabilities of Gatsby which is far more than someone new to
Gatsby wants to see. So I moved the old default starter to
gatsby-starter-kitchen-sink and replaced it with one much more minimal
that'll be much easier to use for people new to Gatsby and just trying
to understand things and to established users of Gatsby who want
something minimal as a base to build out a new site.