fix search box
Text input in search box is retained between page navigations and when you click on the search box with an existing input text and hit Enter, nothing happens. This change fixes this behavious by popping up the search result box on focus of search field and automatically selecting the first result so that when we hit Enter, it opens up the first result.