chore(www): Correct Link to Create a Starter Stub (#12156)
## Description

The `Create a Starter` link in the docs sidebar was giving a 404 page instead of directing to the stub.

To correct this I updated the link to point to the `contributing` directory instead of the `docs` directory in the `doc-links.yaml` file. An alternative solution would be to move `create-a-starter.md` from `/contributing` to `/docs`.