Update/fix to "Creating Tags Pages" documentation (#12470)
<!--
  Have any questions? Check out the contributing docs at https://gatsby.dev/contribute, or
  ask in this Pull Request and a Gatsby maintainer will be happy to help :)
-->

## Description

This updates the markup provided in this documentation (and one minor text tweak to clarify markup change) so that it is not dependent on a GraphQL query to frontmatter.path. This query fails, and @jlengstorf mentioned that this was a bug. I was working through this problem from the gatsby-starter-blog codebase.

Replaced calls to `path` with `slug` and updated the code examples provided (having worked through relevant issues and have tags & categories pages working at https://github.com/shinytoyrobots/shinytoyrobots-dotcom/pull/5).

In addition removed line `filter: { frontmatter: { published: { ne: false } } }` from GraphQL query in tutorial section on creating a Tags index page. This only threw errors and compile failure for me, and this specific documentation page does not appear to be dependent on any other previous tutorial which has required a `published` field, so it seemed unnecessary that it be there.

## Related Issues

<!--
  Link to the issue that is fixed by this PR (if there is one)
  e.g. Fixes #1234, Addresses #1234, Related to #1234, etc.
-->