chore(docs): tweak unit testing to remove invalid props from Link (#12445)
## Description
I followed the Unit-Testing docs to set up jest with Gatsby, and I ran into the following error related to Gatsby's `Link` component:
```
React does not recognize the `activeStyle` prop on a DOM element.
```

## Changes
Removing the `activeStyle` prop from the `Link` mock fixed the issue for me. Seems like this should be the default setup. I was unable to find any discussion of this issue on the internet, sorry if there's something I'm missing.