fix(gatsby-plugin-sharp): strip non related args when hashing resizing args (#12129)
So continuing on the work to make gatsby-plugin-sharp friendly on third-party packages (https://github.com/oorestisime/gatsby-remark-rehype-images/issues/2) we now face this issue https://github.com/gatsbyjs/gatsby/issues/12080

TL;DR images generated from transformer-sharp or and other plugins are the same (same sha256) but with different hash (depends on the available options). The problem is that all the plugins leak their options into the hash generation and even ones that are unrelated to the resizing processing (base64, generateTraceSVG, traceSVG etc)

I've been discussing this with @pieh for the last 2-3 hours :D
Proposed solution was to sanitize the args based on a whitelist, which is implemented on the PR. Which leads to the first question:

- is it ok if after this release _all_ image hashes are changing? 

By doing so it can also potentially help  lazy image processing @wardpeet by getting rid of unrelated options when persisting jobs.

Anyway current state of the PR is that it kind of works, i ve tested this in websites using my plugin and other that not. The issue here is that there are some differences in the number of images created when this modification is on or not.

I also verified that public folder doesn't contain duplicate images:

`find public -regex ".*\.\(jpg\|webp\|png\|jpeg\)" -type f -exec sha256sum {} \; | cut -d ' ' -f 1 |sort| uniq -d | wc -l`