Fix blur-up preview image not working on initial page load (#14816)
Previously this effect only worked for AJAX-based page loads. This change allows the effect to (mostly) work on initial load when the JS isn't yet loaded. If JS is already loaded then the original behaviour is preserved.

This increases the head size by 160 bytes but decreases the size of each image element by 103 bytes.

The head is only downloaded on initial page load (not AJAX loads), so if the visitor encounters 2 or more images during their visit then this change was worth it.