chore(docs): Update the link to global css example (#27250)
* Update the link to default starter index.js 

On line 13, the link to index.js points to a commit from 2018.  The link is there to demonstrate how a Layout component is implemented, and this implementation has changed from 2018 to 2020.  For me, the way it's done in the latest commit is much easier to understand.

Not sure if you want to link to the commit on May 28th, 2020: https://github.com/gatsbyjs/gatsby-starter-default/blob/e1e2a960f8a54e2a1234d1694d390b1cc26b7947/src/pages/index.js#L9

Or just link to the latest commit: https://github.com/gatsbyjs/gatsby-starter-default/blob/master/src/pages/index.js#L9

In the proposed change, I updated to the May 28th commit to be safe.

* Update docs/docs/global-css.md

Co-authored-by: Lennart <lekoarts@gmail.com>

Co-authored-by: Lennart <lekoarts@gmail.com>